Add keywords; improve error messages.
Added an error production which uses END_OF_LINE as a synchronizing
token. This works well for an interactive calculator--for the real
parser, we'll probably use right parens and semicolons as synchronizing
tokens instead.

Added a keyword, "blah", during symbol table initialization. Now we can
perform a specific semantic check: make sure a keyword is not used as an
lvalue.